rentomatic/requirements/test.txt
-r prod.txt
tox
coverage
pytest
pytest-cov
pytest-flask

pytest-docker
docker-compose
pyyaml
psycopg2
sqlalchemy_utils